Compute vb.name automatically using directory name
This change makes it easy to test pull requests using `git worktree` so
long as the working directory is named in the form of
`picoCTF{{suffix}}`. For example, this PR is issued from a branch in a
working directory created using `git worktree add -b
Vagrantfile-auto-name-pr ../picoCTF-Vagrantfile-auto-name-pr`. With this
change, `vagrant up` will automatically assign the machine names to have
a suffix of `-Vagrantfile-auto-name-pr`, which we may reasonably expect
to be unique on the host.