Accept both intended flags and DEBUG_KEY
The old behavior does not accept the intended flag if DEBUG_KEY is set.
That behavior makes testing less smooth because it requires restarting
the web service to switch between accepting DEBUG_KEY for platform
testing and accepting the intended flags for problem testing. This patch
makes both types of testing possible with one VM.