Fix empty-form team registration
The team Join/Register form has a one-off bug where the default
HTML form `action` is to join, and the secondary the Register
function just calls a POST with params. The latter does not go
through form validation, so teamname or teampass could be omitted
as key, causing a nonspecific server error. This is now validated
client-side. This form is present on front-page (after reg) as well
as the account page.