Add shell_server sharding by problems instances
Shell servers are tagged with `server_number` upon add,
and their respective problems are tagged with same.

Teams are assigned `server_numbers` based on configured
manual and auto stepping. Subsequent problem instances
assignment is limited to those problem instances on the
teamn's assigned shell server.

Teams that are assigned to not-yet added shell servers
cannot be assigned problem instances and receive an error
to contact admin.