Add optional docker configuration options and adjust ports
Simplify the blocked port range on the low end, and cap prior to the
ephemeral ports. This prevents a conflict with the ports that docker
might choose when ephemerally mapping containers on-demand.

This should not be a problem (e.g. the ports would already be bound) but
this ensures there is no conflict.

Docker configuration options are entirely optional. Will fall back to
discovering docker configurations from the env if none are specified.